refactor(#625): move VFS hook impls from core to service directories
Delete core/vfs_hook_impls.py and distribute hooks to their owning
service layers:
- DynamicViewerReadHook → services/rebac/dynamic_viewer_hook.py
- AutoParseWriteHook → parsers/auto_parse_hook.py
- TigerCacheRenameHook → bricks/rebac/cache/tiger/rename_hook.py

Fix DynamicViewerReadHook get_subject type: str → tuple[str, str]
to match actual _get_subject_from_context return type.

Fix orchestrator.py mypy errors: remove stale AuditConfig,
event_log, scheduler_service references.

Move test file from tests/unit/core/ to tests/unit/services/.
